+++ src/main/java/com/zeal/controller/PictureController.java
@@ -2,11 +2,10 @@ package com.zeal.controller;
 
 import com.zeal.http.response.Response;
 import com.zeal.service.PictureService;
-import com.zeal.utils.ConfigureUtils;
 import com.zeal.utils.SessionUtils;
-import com.zeal.vo.album.PictureVO;
+import com.zeal.vo.album.AlbumVO;
+import com.zeal.vo.user.UserInfoVO;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,9 +15,6 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.OutputStream;
 
 /**
  * Created by Administrator on 6/29/2016.
@@ -26,51 +22,31 @@ import java.io.OutputStream;
 
 @RequestMapping("/picture")
 @Controller
-public class PictureController {
+public class PictureController extends AbstractController {
 
     @Autowired
     private PictureService pictureService;
 
     @RequestMapping(value = "/{id}")
-    public void find(@PathVariable("id") long id, HttpServletResponse response) {
-        File file = pictureService.getDiskFile(id);
-        if (file.exists() && file.canRead()) {
-            FileInputStream fileInputStream = null;
-            OutputStream stream = null;
-            try {
-                fileInputStream = new FileInputStream(file);
-                byte[] data = new byte[(int) file.length()];
-                fileInputStream.read(data);
-                response.setContentType("image/jpeg");
-                stream = response.getOutputStream();
-                stream.write(data);
-                stream.flush();
-            } catch (IOException e) {
-                e.printStackTrace();
-            } finally {
-                if (fileInputStream != null) {
-                    try {
-                        fileInputStream.close();
-                    } catch (IOException e) {
-                        e.printStackTrace();
-                    }
-                }
-                if (stream != null) {
-                    try {
-                        stream.close();
-                    } catch (IOException e) {
-                        e.printStackTrace();
-                    }
-                }
+    public void find(@PathVariable("id") long id, HttpServletRequest request, HttpServletResponse response) {
+        AlbumVO albumVO = pictureService.findAlbumByPictureId(id);
+        if (!albumVO.isPublished()) {
+            UserInfoVO userInfoVO = SessionUtils.getUserInfo(request);
+            long userId = 0L;
+            if (userInfoVO != null) {
+                userId = userInfoVO.getId();
             }
-
+            pictureService.checkAuthority(id, userId);
         }
+        File file = pictureService.getDiskFile(id);
+        returnFile(response, file);
     }
 
 
     @RequestMapping(value = "/delete/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
     public Response delete(@PathVariable("id") long id, HttpServletRequest httpServletRequest) {
+        pictureService.checkAuthority(id, SessionUtils.getUserInfo(httpServletRequest).getId());
         pictureService.delete(id, SessionUtils.getUserInfo(httpServletRequest).getId());
         return new Response.Builder().success().build();
     }
